dt-bindings: soc: starfive: Add vout-subsystem IP block
Add the dt-binding documentation for the StarFive JH7110 Video Output
(VOUT) subsystem.

This node acts as a parent for all devices within the VOUT power domain,
including the DC8200 display controller, the VOUTCRG clock generator,
and the HDMI MFD block. Its driver is responsible for managing the
shared power domain and top-level bus clocks for these children.

It is a bit similar to the display subsystem qcom,sdm845-mdss DT node.

Signed-off-by: Michal Wilczynski <m.wilczynski@samsung.com>